Change gray_avg to WORD

+++ problems/2022/x/filter-more/helpers.c
@@ -9,7 +9,7 @@
 void grayscale(int height, int width, RGBTRIPLE image[height][width])
 {
     RGBTRIPLE *pixel = NULL;
-    BYTE gray_avg;
+    WORD gray_avg;
     for (int i = 0; i < height; i++)
     {
         for (int j = 0; j < width; j++)
